Move the context file to a separate folder
